requirements.txt
python-telegram-bot==22.0
groq==0.36.0